admin manages users webpage
admin can view all the users here. Once admin clicks the delete button,
the corresponding user is deleted. If "Edit" button is clicked, user
info is put in the "Add/Edit" user part which is on the same page for
updating user info. In the "Add/Edit" part, if user id is not existed, a
new user account is inserted into the database.